Finishing up incorrect behaviour of the converter valve setting
