+++ angelsrefining/locale/en/ore-refining.cfg
@@ -280,6 +280,7 @@ slag-processing-nuc=Crystallizing to Nucular
 
 [item-group-name]
 resource-refining=Resource Refining
+angels-fluid-converter=Converter fluids
 angels-void=Voids
 
 [technology-name]
